+++ src/main/java/com/umc/finly/domain/market/infra/CnnFearGreedIndexProvider.java
@@ -2,7 +2,7 @@ package com.umc.finly.domain.market.infra;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.umc.finly.domain.market.dto.FearGreedResDTO;
+import com.umc.finly.domain.market.dto.res.FearGreedResDTO;
 import com.umc.finly.domain.market.enums.FearGreedStatus;
 import com.umc.finly.domain.market.exception.code.MarketErrorCode;
 import com.umc.finly.global.apiPayload.exception.CustomException;
