+++ src/main/java/com/umc/finly/domain/market/infra/FearGreedIndexProvider.java
@@ -1,6 +1,6 @@
 package com.umc.finly.domain.market.infra;
 
-import com.umc.finly.domain.market.dto.FearGreedResDTO;
+import com.umc.finly.domain.market.dto.res.FearGreedResDTO;
 
 public interface FearGreedIndexProvider {
     FearGreedResDTO getFearGreedIndex();
